% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplification.R
\name{get.data.locations}
\alias{get.data.locations}
\title{Get the list of data locations from an object of class \code{RcppSimplification}.}
\usage{
get.data.locations(x)
}
\arguments{
\item{x}{An object of class \code{RcppSimplification}.}
}
\value{
A #data-by-3 matrix storing the coordinates of data locations.
}
\description{
Get the list of data locations from an object of class \code{RcppSimplification}.
}
\seealso{
\code{\link{simplify.mesh.2.5D}}
}
