% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplification.R
\name{get.edges}
\alias{get.edges}
\title{Get the list of edges from an object of class \code{RcppSimplification}.}
\usage{
get.edges(x)
}
\arguments{
\item{x}{An object of class \code{RcppSimplification}.}
}
\value{
A #edges-by-2 matrix, where the \eqn{i}-th row stores the identifiers of the vertices at the end-points of the \eqn{i}-th edge.
}
\description{
Get the list of edges from an object of class \code{RcppSimplification}.
}
\seealso{
\code{\link{simplify.mesh.2.5D}}
}
